WIP attempt to make a generic function to extract a User ID from an identifier dict.
Unfortunately this is a bit tricky as I'm taking the code from login, but it wants to do login-specific stuff inside of it. The hardest one
is ratelimiting. Ideally we could just pass a ratelimiter object, but all the state isn't kept inside this object.

Going to try and make a PR for that. Could make a great case that this shouldn't be read from the config each time. Perf win!